# rwa/rwa-backend/apps/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "RWA API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/trade/asset": {
            "get": {
                "description": "Get a single asset information by symbol",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Trade"
                ],
                "summary": "Get Asset",
                "parameters": [
                    {
                        "type": "string",
                        "example": "AAPL",
                        "description": "Asset symbol",
                        "name": "symbol",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.GetAssetResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/web.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.Response"
                        }
                    }
                }
            }
        },
        "/trade/assets": {
            "get": {
                "description": "Get list of assets (stocks, crypto, etc.) with filtering options",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Trade"
                ],
                "summary": "Get Assets",
                "parameters": [
                    {
                        "type": "string",
                        "example": "active",
                        "description": "Asset status filter (active or inactive)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "us_equity",
                        "description": "Asset class filter (us_equity or crypto)",
                        "name": "asset_class",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "NASDAQ",
                        "description": "Exchange name filter",
                        "name": "exchange",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.GetAssetsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.Response"
                        }
                    }
                }
            }
        },
        "/trade/currentPrice": {
            "get": {
                "description": "Get the current price for a given symbol",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Trade"
                ],
                "summary": "Get Current Price",
                "parameters": [
                    {
                        "type": "string",
                        "default": "AAPL",
                        "example": "AAPL",
                        "description": "Stock symbol",
                        "name": "symbol",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.GetCurrentPriceResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.Response"
                        }
                    }
                }
            }
        },
        "/trade/historicalData": {
            "get": {
                "description": "Get historical price data for a given symbol",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Trade"
                ],
                "summary": "Get Historical Data",
                "parameters": [
                    {
                        "type": "string",
                        "default": "AAPL",
                        "example": "AAPL",
                        "description": "Stock symbol",
                        "name": "symbol",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1704067200,
                        "example": 1704067200,
                        "description": "Start time as Unix timestamp in seconds",
                        "name": "start_time",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1706745599,
                        "example": 1706745599,
                        "description": "End time as Unix timestamp in seconds",
                        "name": "end_time",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "1d",
                        "example": "1d",
                        "description": "Time interval. Short format: 1m, 5m, 15m, 1h, 1d, 1w. Full format: 1Min, 5Min, 15Min, 1Hour, 1Day, 1Week",
                        "name": "interval",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "example": 100,
                        "description": "Maximum number of records to return",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.GetHistoricalDataListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.Response"
                        }
                    }
                }
            }
        },
        "/trade/latestQuote": {
            "get": {
                "description": "Get the latest bid/ask quote for a given symbol",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Trade"
                ],
                "summary": "Get Latest Quote",
                "parameters": [
                    {
                        "type": "string",
                        "default": "AAPL",
                        "example": "AAPL",
                        "description": "Stock symbol",
                        "name": "symbol",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.GetLatestQuoteResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.Response"
                        }
                    }
                }
            }
        },
        "/trade/marketClock": {
            "get": {
                "description": "Get the current market clock status including whether the market is open and next open/close times",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Trade"
                ],
                "summary": "Get Market Clock",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.GetMarketClockResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.GetAssetResponse": {
            "type": "object",
            "properties": {
                "asset": {
                    "$ref": "#/definitions/dto.AssetDTO"
                }
            }
        },
        "dto.GetAssetsResponse": {
            "type": "object",
            "properties": {
                "assets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.AssetDTO"
                    }
                }
            }
        },
        "dto.GetCurrentPriceResponse": {
            "type": "object",
            "properties": {
                "price": {
                    "type": "number"
                },
                "symbol": {
                    "type": "string"
                },
                "timestamp": {
                    "description": "Unix timestamp in seconds",
                    "type": "integer",
                    "example": 1704067200
                },
                "volume": {
                    "type": "number"
                }
            }
        },
        "dto.GetHistoricalDataListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.GetHistoricalDataResponse"
                    }
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
...
        "dto.GetLatestQuoteResponse": {
            "type": "object",
            "properties": {
                "quote": {
                    "$ref": "#/definitions/dto.QuoteDTO"
                }
            }
        },
        "dto.GetMarketClockResponse": {
            "type": "object",
            "properties": {
                "is_open": {
                    "type": "boolean"
                },
                "next_close": {
                    "description": "Unix timestamp in seconds",
                    "type": "integer",
                    "example": 1704067200
                },
                "next_open": {
                    "description": "Unix timestamp in seconds",
                    "type": "integer",
                    "example": 1704067200
                },
                "timestamp": {
                    "description": "Unix timestamp in seconds",
                    "type": "integer",
                    "example": 1704067200
                }
            }
        },
...
        "web.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Rwa common-api",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}